Simple but working web interface
I currently have a simple yet working web interface for my application.

The best thing to do is to start it in `nohup`, so
`nohup sudo flask --ap my_first_flask run --host=0.0.0.0 --debug`
Then you will be able to access the file at $DEVICE_IP$:5000

Flask warns you that you shouldn't use the --debug flag in
production (it allows me to access it from the lan), but given
the small usage factor of this app, with basically only me and
friends on the network I think it might be fine.