Huge unordered push
I have not done anything with the git branch, so uhm,
here is everything which has been done in the last while

The current state of affairs is that I have two seperate processes,
- The one located in /src/lights, which is a daemon taking netwrok
requests and controlling the lights based uppon those commands
- The other is in src/website, which handles the weblogic,
aka the user io for the system.

Both have their own activation script, and there is a top level
script activating both.